New Plots and New code for additional groupings
Did some data manipulation to make groupings for cod, capelin and the seabirds. See extensive notes on what I did to determine which birds count as 'seabirds', obviously this can be changed as you see fit, but there are two different bird ones, for reasons explained in the code. Overall: This will allow for heatmaps with capelin, cod and seabirds